+++ src/main/java/com/otto/cellescanner/MassiveOsFreakyAddons.java
@@ -20,7 +20,7 @@ public class MassiveOsFreakyAddons {
     // the display name is the new hub brand. See MassiveoAddons.
     public static final String MODID = "cellescanner";
     public static final String NAME = "Massiveo's Freaky Addons 4.2.3";
-    public static final String VERSION = "4.7.1";
+    public static final String VERSION = "4.8.0";
 
     public static CelleConfig config;
     public static CelleScanner scanner;
@@ -95,6 +95,7 @@ public class MassiveOsFreakyAddons {
 
         MinecraftForge.EVENT_BUS.register(new AntiAfk());
         MinecraftForge.EVENT_BUS.register(new BandeEsp());
+        MinecraftForge.EVENT_BUS.register(new PlayerEsp());
         MinecraftForge.EVENT_BUS.register(new ChestAlarm());
         MinecraftForge.EVENT_BUS.register(new ArmorSkins());
         MinecraftForge.EVENT_BUS.register(AutoTrash.INSTANCE);